UPDATE: VM detector & Debug detector
